// src/common/enum/user-type.enum.ts
export enum UserTypeENUM {
  planner = 'asPlanner',
  vendor = 'asVendor',
}

export enum InterestEnum {
  Vendor = 'Vendor',
  venueOwner = 'Venue Owner',
}

export enum Role {
  Admin = 'Admin',
  Planner = 'Planner',
  Moderator = 'Moderator',
  Vendor = 'Vendor',
}

export enum VenueTypesEnum {
  Conference_Center = 'Conference Center',
  Sport_Area_Stadium = 'Sport Area/Stadium',
  Cultural_Center = 'Cultural Center',
  Hotel_Resort = 'Hotel and Resort',
  Banquet_Hall = 'Banquet Hall',
}

export enum VendorTypesEnum {
  Mc = 'Mc',
  Event_Planner = 'Event Planner',
  Interior_Decoration = 'Interior Decoration',
  Photography = 'Photography',
  Catering = 'Catering',
}

export enum GenderEnum {
  Female = 'Female',
  Male = 'Male',
  Unspecified = 'Unspecified',
}